rag_requirements.txt
# === Существующие зависимости из основного проекта ===
openai>=1.12.0
tqdm>=4.66.0
tenacity>=8.2.3
openpyxl>=3.1.0
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv>=1.0.0

# === Новые зависимости для RAG системы ===

# LangChain и компоненты
langchain>=0.1.0
langchain-community>=0.0.20

# Векторная БД Qdrant
qdrant-client>=1.7.0

# Embeddings модель
sentence-transformers>=2.3.0

# PyTorch (для sentence-transformers)
# Для CPU (без GPU):
torch>=2.1.0
# Если есть GPU, установите соответствующую версию:
# torch>=2.1.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# Дополнительные утилиты
numpy>=1.24.0
scikit-learn>=1.3.0

# Для тестирования (опционально)
pytest>=7.4.0
pytest-benchmark>=4.0.0

# Для SSE клиента (опционально, для примеров интеграции)
sseclient-py>=1.8.0